fix: remove FLAG_GRANT_READ_URI_PERMISSION in gallery helper
Given the limit to persistable permissions for each app on some devices, and considering that Raccoon does not need to perform long tasks on the file but just a one-shot read to upload it to the server, the flag can be removed. This will solve some crashes that have been reported by end users.